requirements.txt
./dependencies/awscli-1.29.21-py3-none-any.whl
./dependencies/boto3-1.28.21-py3-none-any.whl
./dependencies/botocore-1.31.21-py3-none-any.whl
faiss-cpu
transformers
Pillow
pypdf
langchain
streamlit
urllib3==1.26.6
stability-sdk[sagemaker] @ git+https://github.com/Stability-AI/stability-sdk.git@sagemaker
python-resize-image
